ACTUALLY fix imitating random users, and more
I wasn't processing the data I was getting from s3listUserIds() correctly. I wasn't accounting for the file extensions and preceding paths that were part of each Key. As part of this change, Bipolar now requires Path.
Also, I noticed a race condition that pretty much always failed in appendCorpus(); the function was almost always resolving before the above call to s3read() finished. I fixed this by putting the resolve() in a .finally() statement, so now it only executes when s3read() is done.
I also added a "save before closing" thing, so that Bipolar will try to save the unsaved files to S3 and clear its cache before exiting when the operating system tells it to close.