Fix imitating random users for real this time
It was a typo. I made a typo. I was renaming a variable and forgot to rename it in one spot and that was what was causing the random user imitation code to fail.
Bipolar now also caches a list of the IDs of the users it knows, so it doesn't have to request that information from S3 every time it needs to imitate a random user. This will cut down on unnecessary S3 requests. Those things cost money, yo.