Remove NAME variable
NAME was only used in two places in the file. Even then, they were both for logging. Turns out the NAME variable is actually used by Linux and was causing problems when run on that operating system. I didn't think it was worth keeping around, so I rewrote those log lines to just not use the environment variable, and now Bipolar no longer uses it.