Add support for imitating by ID for real
I just learned the hard way that parseInt() can be imprecise when working with as many as 18 digits. Sometimes, the ID you enter would not be the ID Bipolar looked for in the system because parseInt() would take in something like "206235904644349953" and output 206235904644349950.

Turns out, though, I didn't even need a parseInt() in this scenario; isNaN() actually works with strings, too ('123' is not NaN; 'foo' is NaN). So I just removed parseInt() and LG lifes good.