Make more functions async
I found Promise-based versions of the s3 functions, so now they're async. All that are left are the node FS functions. Still need to figure out how to the use the Promise versions of them.
Added some dumb code to log.imitate to support plain text mode.
Fixed a typo in a log message lol