Fix handleCommands() typo
Looks like there was one holdover from when this function was Promise-based that I forgot to change. The check to make sure the author isn't a bot was accidentally set to resolve instead of return. This is fixed so now Schism won't DM the admins an error if a bot tries to give her a command.