randomUserID() only returns users that Schism can find
If a user doesn't exist (as far as Schism knows), that user will no longer be picked by randomUserID(). Before, it would pick directly from the list of IDs in the corpus directory. This would sometimes result in the function picking a user that isn't actually in any server with Schism, resulting in the less-pretty fallback imitate message with the userID for name and the default avatar. Now, randomUserID() only picks users Schism can actually access. You can still imitate users that Schism has a corpus for; it's just now Schism will never pick one of those users on her own.

This change also resulted in randomUserID() becoming async. The places that called randomUserID() have been updated accordingly.

Also, on-message logic has been improved yet again, now making sure to never do anything at all if the message does not have any text content (this can happen if the message is just an image, for example). On top of that, Schism now does not learn if the message starts with the command prefix, in order to prevent her from putting a bunch of "|imitate [user]"'s in her corpi. This check was actually present a long time ago but it must have gotten lost sometime throughout this project's life.