Fix scrape() inserting "undefined"
scrape() was putting "undefined" at the beginnings of files because it was appending to `undefined` every first time it added text to a scrapeBuffer. A check has been put in place to make sure that, if the buffer of text being built up for a user does not exist, create it before appending any data.

Also, scrape() now filters Schism commands from the data it adds. No more "|imitate me" all over people's corpi.