scrape() revert to polling
Promise.all() wasn't working for properly detecting when the recursive function _getBatchOfMessages() was fully done executing. I have reverted scrape() back to the old and ugly code that I had before.
Instead of `async/await`, this function now uses `return new Promise()` syntax again.
Instead of pushing all the instances of _getBatchOfMessages() to an array that is Promise.all()'d, scrape() will poll for the variable `processes` to be 0, which only occurs when _getBatchOfMessages() is fully done executing (or the race condition fails rofl).