Fix scraping
local has been renamed to onS3, because I had confused myself and that is actually what it meant this whole time. 
The place where userIDs get added to this variable is different now, where that ID gets data written to S3 instead of where the ID gets data written to cache. This will probably fix scraping.