Fix typo
I forgot to add `throw` before a string. As a result, where the code was supposed to be throwing an error, it would instead just see a random string floating around, be like, "ok" and move on without doing anything. This code was a change I made recently to help me track down the "Unknown user" error, so maybe NOW it will do its job.