Disable @everyone
I bet you can guess what happened.
It turns out that, even when @everyone is disabled on a server, Webhooks still _can_ ping @everyone. A simple way to keep this from happening is to add the disableEveryone option to the WebhookClient constructor found in parseHooksDict(). I also added it to Schism's main DiscordClient for good measure.