Do not randomly choose IDs from other servers
randomUserID() has been changed so that it can no longer randomly choose a user outside of the server the code calling randomUserID() was triggered from. This is a small step toward getting Schism ready for public use.

randomUserID() now requires a guild argument so that it can only return user IDs that belong to a member of that server. It is up to the code that calls randomUserID() to give it the server that the code was triggered by.

Also, markov.randomUserID() and markov._elementAt() were moved to this file since it made no sense for them to be in markov.js. _elementAt() was renamed to just elementAt() to reflect that it is no longer private to markov.js. As a result, markov.js now returns just generateSentence() instead of an object containing multiple methods. index.js has been refactored accordingly.