Fix imitating in DMs
I discovered that a recent change made it so that Schism couldn't imitate in DMs. This was caused by how imitate() always tried to use .fetchMember() (a method only available for guilds) even when the channel the command came from is not a part of a guild, like a DM channel. This change makes it so that imitate() will only use .fetchMember() when the channel the command came from is a part of a guild. Otherwise, imitate() will use .fetchUser().

This change also fixes a minor flaw there the term "Not " would appear in front of the name in places where it shouldn't like in the logs and in text mode. "Not user" should only appear as the name of the Webhook, and as just "user" everywhere else. This was fixed by only adding "Not " to the name when Schism is outputting through a Webhook.