Fix capitalization in |intimidate
Last change fixed a bug keeping Schism from imitating in DMs, but it also made it so that the "Not " in front of people's names in Webhook mode stopped being capitalized like it should be while intimidating. This change fixes that by making sure to also capitalize the "Not " while imitate() is in intimidate mode.