Create expirable-set.js
This is a class that adds a new method to the ES6 Set. This method adds a value, then deletes it after a set (haha) period of time. Schism uses it to keep track of pending "|forget" command confirmations and to let them expire after the confirmation window has passed.