EULA- and forget-related functions
The S3 helper module can now remove files! Took long enough. Schism uses this to forget users.
Also, there are now more the ceremonious readFile() and writeFile() to go along with the simplier read() and write() for when you want to read or write something that isn't strictly a corpus file in a specific directory. Schism uses this to store the file that keeps track of who has agreed to the EULA.