Create tracker-set.js
This class augments the add() and delete() methods of the ES6 Set such that it can remember with a Boolean value if those methods have ever been used. Of course, you can just change .modified back to false to reset its state. Schism uses this kind of Set to know if it needs to push a modified version of the EULA-consent file to the Source of Truth when it comes time to save().