Rename Schism to Parrot
After doing some thinking, I've decided that Parrot is a more fitting name for this bot.
If you look back in the beginnings of this repository, you will see that before the bot was named Schism, it was named Bipolar. I wanted a name that represented how the bot has "multiple personalities". But then it came to my attention that bipolar disorder is not actually about having multiple personalities. And it's also really uncool to name something after a mental illness. Real cringe moment for me.
Still fixated on how there are multiple parts of the bot, I tried to come up with some cool word that means "divide", and I decided on Schism.
That was the bot's name for a long time, but I couldn't help but feel like it was sort of a bad name since the meaning behind it had ventured so far away from the bot's original purpose.
So I shifted my focus on the bot's purpose away from "divide" and onto "imitate"... then the name became obvious.
From here on out, the bot that imitates everyone is named Parrot.