Semicolons!!
I put the semicolons back where they arguably belong.
I have also changed where I put the commas in dictionaries from the beginning of the entry to the end of it. Also, there are now trailing commas on the last entry of each dictionary. Out with one questionable JavaScript formatting choice and in with a new one, right?