I need to start doing real commits
All the commands but speaking and learning have been tested and they work at this point! Need to test those other two commands.

Next step is implementing learning from messages. With other things referring to the feature and the config file filled in for it, the code could hardly be more ready for it.

Embeds have all been consolidated to a single file. Every new Message Embed will be put in this file and imported wherever.

All FS calls have been changed to be async. I wanted to fool myself into thinking that I could use synchronous for this, but I didn't. A LOT of methods have turned from Blue to Red as a result of this change. (see https://journal.stuffwithstuff.com/2015/02/01/what-color-is-your-function/)

I made some basic types. They're in src/index.d.ts.
Also, all of the interfaces that represent data structures, like JSON files, are gonna go here, too. Ones for Corpus and Config are already there.