Further progress
Getting close!

Today I Learned that Discord.MessageCollector does not run through a channel's history, but actually passively listens for messages in a channel. That is not what I want.
So I made my own channel crawler! I reused a lot of code for this part from Parrot 1, except making it look a lot better with the experience I gained in JavaScript since I wrote the original. I gave it the closest interface I could to Discord.MessageCollector, because they really do a similar task, just mine on a faster scale. And also maybe because I had already written code to interface with Discord.MessageCollector *cough*.
I almost want to publish my Crawler to NPM as a standalone unit. That's on my todo list now.

I modified the imitate command (why did I do it _this_ way, though?) to wrap it in a pair of start- and stopTyping() statements. This command could take a long time to come up with its result, so I figured it deserves that typing indicator. This would have been easier to do in discord.py, where the stopTyping() call is handled for you.

Learning is finally here! And to my surprise, it works! Or at least it works at its core; I can't say for sure about the console funny-business I added on top of it, since I haven't tested that part yet.

Miscellanous stuff:
  - More fun utility functions!
  - More comments!
  - "ayy lmao" support is back!

Much to my own surprise, Parrot's overhaul is almost complete already! At this rate, it could just be a few more days before its core functionality is all complete, then all that's left to do is contribute to its feature bloat for the rest of time!