Fix Quickstart some but not all the way
The Channel Crawler was ending immediately because I had a couple conditions switched around the wrong way. That is fixed now. Quickstart is still giving the wrong numbers, though.

New embed message for when your corpus is corrupt. Obviously, I hope this never actually happens in production, but it happened in development, so I guess I might as well have an at least half-friendly way to tell it to users than an error message.

Now that ParrotEmbed exists, it's no longer necessary to specify values that happen to be one of ParrotEmbed's default values. As such, the embeds in this file no longer explicity state to make their color purple, because ParrotEmbed does that by default.

The Learning module;'s learnFromMany has been merged with learnFrom. It is now one function that takes in either a Message or an array of Messages. Just make sure the arrays you give it all have the same author.