Support adding channels without a webhook
Parrot was already supposed to be able to do this, but I forgot to make part of the code support it. You should now be able to specify `null` in place of a WebhookConfig in the config.ts file and Parrot will use text mode in that channel instead of a Webhook.