Allow imitating unregistered bots
Now you can have Parrot imitate bots even when the bot isn't registered (because you can't register a bot).

This requires more data about the User than just their ID, so this change made it so that CorpusManager now takes in a User in all of its methods in every place that it used to take in a User ID.

Every use of a CorpusManager method has been updated accordingly.