Python rewrite
I'm throwing discord.js in the trash. I'm tired of it. This time, Parrot's gonna be Python.

A new addition already is the vocodes command. Parrot can now join a voice channel and interface with vo.codes to imitate speaker from vo.codes's library of voices, through Discord.

I'm also making a point to minimize the number of things you have to configure through a file, instead trying to make as many things as possible configurable through commands. Already, you can configure learning and speaking channels through a command, and (with the right channel permissions) Parrot will create Webhooks on its own as needed.

I aim to push this to main as soon as the imitate command is stable. It might still be missing features that the old version had at that point, but at least it won't crash every four hours.