Switch to aiohttp
A synchonrous requests API is absolutely awful for what Parrot's using it for in this command. For the whole 5-30+ seconds it's waiting for its audio data, Parrot is frozen(!) and will not respond to anything, because the request is blocking the whole thread.

A quick switch to aiohttp, and this is no longer a problem. Parrot now makes the vo.codes request asynchronously so it can keep doing other things while waiting.

This new way of doing things also made it so that this file doesn't import atexit anymore to close the HTTP session it used to have open.

Also, the embed is now properly edited into a kind of "failure" state when the request fails, as opposed to just staying on the loading state as it did before.

Character limit for vocodes sentences has also been bumped up from 280 to 1000. 280 is the limit the vo.codes website gives, but it turns out it will happily accept longer sentences. The only reason I'm giving it a limit at all is because the request tends to time out if the sentence is too long.