FriendlyError bugfix
I pushed code without testing it! Garlic moment. Turns out, as I had it before, Parrot would print an error to console every time a non-registered user would speak, as a byproduct of some part of the learning code generating a FriendlyError. The problem is that FriendlyError was broken because it tried to modify a tuple.

FriendlyError is now fixed by not modifying tuples.