Fix annoying typing issue
mypy really doesn't like when I try to convert something that may or may not be a list into something that is definitely a list. I had to introduce some weird cast() thing to fix it.

Also, I refactored the checks within validate_message to be less ugly.

Does the multi part string on line 89 need a backslash at the end? I think it does... oh well, it has one now.