Fix and comment Userlike
Userlike wasn't working because Guild.get_members() inexplicably isn't working no matter what I try. So I refactored it to use Guild.fetch_members(), instead, which the docs say you shouldn't do. Well, docs, I wouldn't do it if the thing you recommended actually worked!

Userlike now works properly in DMChannels. Before, it didn't work because it tried to call a method on ctx.guild, which doesn't exist in DMChannels.

Now, if you pass in something that doesn't parse to an int, it will just say the user isn't found instead of throwing a spooky ValueError.

I also changed the indentation in the bulleted list in the docstring to use four spaces so that the indentation width doesn't get screwed up throughout in the file in VS Code (and probably other IDEs).