Switch to file.io and aiohttp
Parrot used to use anonymousfiles.io to host corpus downloads pretty much because it was the only file hosting service I knew of off the top of my head that had a free API. But recently, I found out about files.io, which looks way less sketchy, so Parrot uses that site now, instead.

Parrot used to use requests, but come on, that's synchronous. This is pretty much just catching up the download command up vocodes, which is already async.