Throw NoDataErrors instead of KeyErrors
I realized that every place that driving code is calling a CorpusManager method that can throw a KeyError, the KeyError is just being caught and turned into a NoDataError. CorpusManager has all the information it needs to do that itself, so I figured it just will.

I also changed some comments slightly and fixed a couple typing-related oversights.