Add FromList subclass
I noticed that I was reusing code in some places to do the same thing with CustomEmbedPaginator. So I created a subclass of CustomEmbedPaginator that does this in just one place. FromList takes in a list of values to put into embed fields and creates a paginated embed for displaying all of those values. The vo.codes category uses this to display the names of the voices it can use, and the channels and quickstart categories use it to list channel names.