Make ParrotMarkov not copy and paste entire sentences
I misunderstood how parsed_sentences works. I'm not using it anymore, because the Markov was basically just choosing an entire sentence at random and passing it off as its own without modifying it at all. That won't do. The way that I have set it up now makes it so that it is more creative.