Wait for DB to finish loading before logging in
If you start the bot and the database at the same time, the bot may log in and be able to make requests to the database before the database has finished loading its dataset into memory. This can be avoided by just not letting the bot log in until the database is ready to receive commands.