New gibberish algorithm
It's about time I did something—anything—to how Parrot makes gibberish, because how it used to do it  is a bruh moment. When I realized that ThinkZone's gibberish generator is pretty much just a character-based Markov chain generator, I tried to make markovify make a model out of characters instead of words, and this is the result. Might improve this a little more in the future.