Learning: no more lists of messages
Allowing lists of messages as input on top of singular messages to the methods responsible for learning turned out to make the methods a lot more complicated than they needed to be. And it turns out nowhere in the code is a list of messages passed into learn_from() anyway, so nothing else even needed to be changed.