Mirror and invert imitate avatars
To accomplish this, Parrot now requires PIL for image manipulation. Since Discord doesn't allow using a file as a webhook avatar, Parrot has to upload the manipulated avatars before it can use them. So now Parrot uses a special server to dump all its modified avatars just to turn around and use the link to the avatar it posts in as the avatar of its webhook. But that's OK, because this facilitates caching the images. Parrot uses a new "avatars" hash in its database to keep track of whether it's processed any given user's current avatar before.