Update the policy and its location
It's just too much to ask the user to read the privacy policy. They will if they want to. This updates the privacy policy in a few ways I think it should be, including making it more accessible for reading from GitHub instead of Discord.
To balance this out, I'm putting a link to the policy in Parrot's bot description.