Simplify the timestamp migration
Following new findings about the actual state of Parrot v1's database, the role of the timestamp converting migration has changed. There are not actually any integer timestamps with any value other than 0. And most timestamps are ISO 8601 date strings. So in spite of the schema defined for the database, most timestamps present in v1 are actually properly formed. There are just a few (hundred) that are malformed. So I changed this migration to instead just fix them, and leave the rest of the timestamps untouched.